Refactor module package and update config
