+++ README.md
@@ -1,11 +1,10 @@
-# Mẫu EzXhepler
-Một mẫu cho [EzXHelper](https://github.com/KyuubiRan/EzXHelper)
+# DexModule Example
+Mẫu Xposed module sử dụng [EzXHelper](https://github.com/KyuubiRan/EzXHelper)
 
 ## Sử dụng mẫu này để tạo dự án của bạn
-Trước khi sử dụng:  
-- Thay đổi `applicationId` và `namespace` trong `build.gradle.kts`  
-- Thay đổi tên gói và cập nhật đường dẫn lớp HookEntry trong `xposed_init`  
-- Thay đổi `rootProject.name` trong `settings.gradle.kts`  
-- Đồng bộ hóa Gradle  
-- Thay đổi `TAG` và `PACKAGE_NAME_HOOKED` trong `MainHook.kt`  
-- Thay đổi `xposedscope` trong `arrays.xml`
+Trước khi sử dụng:
+- Thay đổi `applicationId` và `namespace` trong `build.gradle.kts`
+- Cập nhật tên gói và đường dẫn `xposed_init`
+- Chỉnh `rootProject.name` trong `settings.gradle.kts`
+- Đồng bộ Gradle
+- Điều chỉnh giá trị trong `MainHook.kt` và `arrays.xml`
