+++ app/src/main/assets/xposed_init
@@ -1,1 +1,1 @@
-com.KTA.Dex.MainHook
+com.example.dexmodule.MainHook
