+++ app/src/main/res/values/arrays.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string-array name="xposedscope">
-    <item>com.android.systemui</item>
+    <item>com.example.targetapp</item>
     </string-array>
 </resources>
